src/requirements.txt
appnope==0.1.2
argcomplete==1.12.3
astroid==2.8.2
backcall==0.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.0
distlib==0.3.3
entrypoints==0.3
filelock==3.4.0
greenlet==1.1.2
idna==3.2
ipykernel==6.5.1
ipython==7.29.0
isort==5.9.3
jedi==0.18.1
jupyter-client==7.0.6
jupyter-core==4.9.1
lazy-object-proxy==1.6.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mysql-connector-python==8.0.27
nest-asyncio==1.5.1
packaging==21.3
parso==0.8.2
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
pipx==0.16.4
platformdirs==2.4.0
playwright==1.15.3
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pyee==8.2.2
Pygments==2.10.0
pylint==2.11.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
selenium==3.141.0
selenium-stealth==1.0.6
six==1.16.0
stevedore==3.5.0
TikTokApi==4.0.4
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
userpath==1.7.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
websockets==10.0
wrapt==1.12.1
youtube-dl==2021.6.6
